test: ExportableFixture for JSON tests exporting
Extract utils for JSON tests exporting to abstract `ExportableFixture`.
